Moved "getSiblingNodes()" util from ListStyleCommand to public utils.
